feat(provider): 添加长桥证券接口并优化期权数据处理
- 新增 SecurityGateway 接口及其实现类 SecurityGatewayImpl
- 添加 OptionsStrategyData 模型类用于存储策略数据
- 修改 OptionsChain 和 Options 模型，增加策略数据相关字段
- 优化期权实时数据处理，使用 BigDecimal 替代 Double
- 新增期权年化收益率计算逻辑

+++ api/src/main/java/me/dingtou/options/model/OptionsExtData.java
@@ -2,6 +2,8 @@ package me.dingtou.options.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 期权扩展数据
  *
@@ -37,7 +39,7 @@ public class OptionsExtData {
     /**
      * 行权价
      */
-    private Double strikePrice;
+    private BigDecimal strikePrice;
     /**
      * 是否停牌
      */
